// typescript/quipu_client/src/apis/AccountingCategoriesApi.ts
export interface GetAccountingCategoriesRequest {
    filterKind?: GetAccountingCategoriesFilterKindEnum;
    filterPrefix?: string | null;
}

export interface GetAccountingCategoryRequest {
    id: string;
}

/**
 * 
 */
export class AccountingCategoriesApi extends runtime.BaseAPI {

    /**
     * Creates request options for getAccountingCategories without sending the request
     */
    async getAccountingCategoriesRequestOpts(requestParameters: GetAccountingCategoriesRequest): Promise<runtime.RequestOpts> {
        const queryParameters: any = {};

        if (requestParameters['filterKind'] != null) {
            queryParameters['filter[kind]'] = requestParameters['filterKind'];
        }

        if (requestParameters['filterPrefix'] != null) {
            queryParameters['filter[prefix]'] = requestParameters['filterPrefix'];
        }

        const headerParameters: runtime.HTTPHeaders = {};

        if (this.configuration && this.configuration.accessToken) {
            // oauth required
            headerParameters["Authorization"] = await this.configuration.accessToken("oauth2", []);
        }


        let urlPath = `/accounting_categories`;

        return {
            path: urlPath,
            method: 'GET',
            headers: headerParameters,
            query: queryParameters,
        };
    }

    /**
     * Retrieve all accounting categories
     */
    async getAccountingCategoriesRaw(requestParameters: GetAccountingCategoriesRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<AccountingCategoryCollection>> {
        const requestOptions = await this.getAccountingCategoriesRequestOpts(requestParameters);
        const response = await this.request(requestOptions, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => AccountingCategoryCollectionFromJSON(jsonValue));
    }

    /**
     * Retrieve all accounting categories
     */
    async getAccountingCategories(requestParameters: GetAccountingCategoriesRequest = {}, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<AccountingCategoryCollection> {
        const response = await this.getAccountingCategoriesRaw(requestParameters, initOverrides);
        return await response.value();
    }

    /**
     * Creates request options for getAccountingCategory without sending the request
     */
    async getAccountingCategoryRequestOpts(requestParameters: GetAccountingCategoryRequest): Promise<runtime.RequestOpts> {
        if (requestParameters['id'] == null) {
            throw new runtime.RequiredError(
                'id',
                'Required parameter "id" was null or undefined when calling getAccountingCategory().'
            );
        }

        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};

        if (this.configuration && this.configuration.accessToken) {
            // oauth required
            headerParameters["Authorization"] = await this.configuration.accessToken("oauth2", []);
        }


        let urlPath = `/accounting_categories/{id}`;
        urlPath = urlPath.replace(`{${"id"}}`, encodeURIComponent(String(requestParameters['id'])));

        return {
            path: urlPath,
            method: 'GET',
            headers: headerParameters,
            query: queryParameters,
        };
    }

    /**
     * Retrieve a specific accounting category
     */
    async getAccountingCategoryRaw(requestParameters: GetAccountingCategoryRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<AccountingCategoryResponse>> {
        const requestOptions = await this.getAccountingCategoryRequestOpts(requestParameters);
        const response = await this.request(requestOptions, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => AccountingCategoryResponseFromJSON(jsonValue));
    }

    /**
     * Retrieve a specific accounting category
     */
    async getAccountingCategory(requestParameters: GetAccountingCategoryRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<AccountingCategoryResponse> {
        const response = await this.getAccountingCategoryRaw(requestParameters, initOverrides);
        return await response.value();
    }

}

/**
 * @export
 */
export const GetAccountingCategoriesFilterKindEnum = {
    Income: 'income',
    Expenses: 'expenses',
    Employee: 'employee',
    Supplier: 'supplier',
    Client: 'client',
    Assets: 'assets'
} as const;
export type GetAccountingCategoriesFilterKindEnum = typeof GetAccountingCategoriesFilterKindEnum[keyof typeof GetAccountingCategoriesFilterKindEnum];