// typescript/quipu_client/src/apis/AuthenticationApi.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Quipu API
 * Quipu is an invoice and taxes software as a service for freelances and companies.  To use our API you need to sign up in our software. You will have 15 days to try it for free. If you need more time, contact our support team to extend the trial period.  For more information about our API, libraries and other applications, visit: https://getquipu.com/en/integrations.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import * as runtime from '../runtime';
import type {
  TokenResponse,
} from '../models/index';
import {
    TokenResponseFromJSON,
    TokenResponseToJSON,
} from '../models/index';

export interface GetAccessTokenRequest {
    grantType: GetAccessTokenGrantTypeEnum;
    scope: string;
}

/**
 * 
 */
export class AuthenticationApi extends runtime.BaseAPI {

    /**
     * Creates request options for getAccessToken without sending the request
     */
    async getAccessTokenRequestOpts(requestParameters: GetAccessTokenRequest): Promise<runtime.RequestOpts> {
        if (requestParameters['grantType'] == null) {
            throw new runtime.RequiredError(
                'grantType',
                'Required parameter "grantType" was null or undefined when calling getAccessToken().'
            );
        }

        if (requestParameters['scope'] == null) {
            throw new runtime.RequiredError(
                'scope',
                'Required parameter "scope" was null or undefined when calling getAccessToken().'
            );
        }

        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};

        const consumes: runtime.Consume[] = [
            { contentType: 'application/x-www-form-urlencoded' },
        ];
        // @ts-ignore: canConsumeForm may be unused
        const canConsumeForm = runtime.canConsumeForm(consumes);

        let formParams: { append(param: string, value: any): any };
        let useForm = false;
        if (useForm) {
            formParams = new FormData();
        } else {
            formParams = new URLSearchParams();
        }

        if (requestParameters['grantType'] != null) {
            formParams.append('grant_type', requestParameters['grantType'] as any);
        }

        if (requestParameters['scope'] != null) {
            formParams.append('scope', requestParameters['scope'] as any);
        }


        let urlPath = `/oauth/token`;

        return {
            path: urlPath,
            method: 'POST',
            headers: headerParameters,
            query: queryParameters,
            body: formParams,
        };
    }

    /**
     * Get an access token
     */
    async getAccessTokenRaw(requestParameters: GetAccessTokenRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<TokenResponse>> {
        const requestOptions = await this.getAccessTokenRequestOpts(requestParameters);
        const response = await this.request(requestOptions, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => TokenResponseFromJSON(jsonValue));
    }

    /**
     * Get an access token
     */
    async getAccessToken(requestParameters: GetAccessTokenRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<TokenResponse> {
        const response = await this.getAccessTokenRaw(requestParameters, initOverrides);
        return await response.value();
    }

}

/**
 * @export
 */
export const GetAccessTokenGrantTypeEnum = {
    ClientCredentials: 'client_credentials'
} as const;
export type GetAccessTokenGrantTypeEnum = typeof GetAccessTokenGrantTypeEnum[keyof typeof GetAccessTokenGrantTypeEnum];