// typescript/quipu_client/src/apis/BookEntriesApi.ts
export interface GetBookEntriesRequest {
    filterType?: GetBookEntriesFilterTypeEnum;
    filterKind?: GetBookEntriesFilterKindEnum;
    filterQ?: string | null;
    filterPeriod?: string | null;
    filterPaymentStatus?: GetBookEntriesFilterPaymentStatusEnum;
    filterContactId?: string | null;
    filterFieldQuery?: GetBookEntriesFilterFieldQueryParameter;
    sort?: string;
    pageNumber?: number;
}

/**
 * 
 */
export class BookEntriesApi extends runtime.BaseAPI {

    /**
     * Creates request options for getBookEntries without sending the request
     */
    async getBookEntriesRequestOpts(requestParameters: GetBookEntriesRequest): Promise<runtime.RequestOpts> {
        const queryParameters: any = {};

        if (requestParameters['filterType'] != null) {
            queryParameters['filter[type]'] = requestParameters['filterType'];
        }

        if (requestParameters['filterKind'] != null) {
            queryParameters['filter[kind]'] = requestParameters['filterKind'];
        }

        if (requestParameters['filterQ'] != null) {
            queryParameters['filter[q]'] = requestParameters['filterQ'];
        }

        if (requestParameters['filterPeriod'] != null) {
            queryParameters['filter[period]'] = requestParameters['filterPeriod'];
        }

        if (requestParameters['filterPaymentStatus'] != null) {
            queryParameters['filter[payment_status]'] = requestParameters['filterPaymentStatus'];
        }

        if (requestParameters['filterContactId'] != null) {
            queryParameters['filter[contact_id]'] = requestParameters['filterContactId'];
        }

        if (requestParameters['filterFieldQuery'] != null) {
            queryParameters['filter[field_query]'] = requestParameters['filterFieldQuery'];
        }

        if (requestParameters['sort'] != null) {
            queryParameters['sort'] = requestParameters['sort'];
        }

        if (requestParameters['pageNumber'] != null) {
            queryParameters['page[number]'] = requestParameters['pageNumber'];
        }

        const headerParameters: runtime.HTTPHeaders = {};

        if (this.configuration && this.configuration.accessToken) {
            // oauth required
            headerParameters["Authorization"] = await this.configuration.accessToken("oauth2", []);
        }


        let urlPath = `/book_entries`;

        return {
            path: urlPath,
            method: 'GET',
            headers: headerParameters,
            query: queryParameters,
        };
    }

    /**
     * Retrieve all invoices, tickets and paysheets
     */
    async getBookEntriesRaw(requestParameters: GetBookEntriesRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<BookEntryCollection>> {
        const requestOptions = await this.getBookEntriesRequestOpts(requestParameters);
        const response = await this.request(requestOptions, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => BookEntryCollectionFromJSON(jsonValue));
    }

    /**
     * Retrieve all invoices, tickets and paysheets
     */
    async getBookEntries(requestParameters: GetBookEntriesRequest = {}, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<BookEntryCollection> {
        const response = await this.getBookEntriesRaw(requestParameters, initOverrides);
        return await response.value();
    }

}

/**
 * @export
 */
export const GetBookEntriesFilterTypeEnum = {
    Invoices: 'invoices',
    Tickets: 'tickets',
    Paysheets: 'paysheets'
} as const;
export type GetBookEntriesFilterTypeEnum = typeof GetBookEntriesFilterTypeEnum[keyof typeof GetBookEntriesFilterTypeEnum];
/**
 * @export
 */
export const GetBookEntriesFilterKindEnum = {
    Income: 'income',
    Expenses: 'expenses',
    Employee: 'employee',
    Supplier: 'supplier',
    Client: 'client',
    Assets: 'assets'
} as const;
export type GetBookEntriesFilterKindEnum = typeof GetBookEntriesFilterKindEnum[keyof typeof GetBookEntriesFilterKindEnum];
/**
 * @export
 */
export const GetBookEntriesFilterPaymentStatusEnum = {
    Paid: 'paid',
    Unpaid: 'unpaid',
    Due: 'due',
    Pending: 'pending'
} as const;
export type GetBookEntriesFilterPaymentStatusEnum = typeof GetBookEntriesFilterPaymentStatusEnum[keyof typeof GetBookEntriesFilterPaymentStatusEnum];